GraphQL optimisation, comments and an addition
GraphQL = Prevent the use of the ANIME MediaType directly, when we are already sending the MediaType after.

Comments = Commented the structures to provide information linked to each part of it.

Additions = Added an "Adults only" information on the footer, to permit the end user to know if the Anime or Manga he's looking at is intended for mature audiences.

Personal addition = Some little English mistakes were fixed, and some errors now get an uppercase on first letter.