Update GraphQL Queries
Reason for that is, since the result of Media is a node linked to the Character, there's no need for us to search for it by popularity after. Of course, the character could be appearing in many versions of the anime/manga, but well, we do not even show it on the bot.